+++ src/Core/TaskCache.cs
@@ -13,7 +13,7 @@ namespace GreenDonut
         private const int _cleanupDelay = 10;
         private readonly ConcurrentDictionary<object, CacheEntry>
             _cache = new ConcurrentDictionary<object, CacheEntry>();
-        private CancellationTokenSource _dispose;
+        private CancellationTokenSource _disposeTokenSource;
         private bool _disposed;
         private readonly LinkedList<object> _ranking =
             new LinkedList<object>();
@@ -128,17 +128,18 @@ namespace GreenDonut
         {
             if (SlidingExpirartion > TimeSpan.Zero)
             {
-                _dispose = new CancellationTokenSource();
+                _disposeTokenSource = new CancellationTokenSource();
 
                 Task.Factory.StartNew(async () =>
                 {
-                    while (!_dispose.Token.IsCancellationRequested)
+                    while (!_disposeTokenSource.Token.IsCancellationRequested)
                     {
                         DateTimeOffset removeAfter = DateTimeOffset.UtcNow
                             .Subtract(SlidingExpirartion);
 
                         if (_ranking.Last != null &&
-                            _cache.TryGetValue(_ranking.Last.Value,
+                            _cache.TryGetValue(
+                                _ranking.Last.Value,
                                 out CacheEntry entry) &&
                             removeAfter > entry.LastTouched)
                         {
@@ -146,8 +147,10 @@ namespace GreenDonut
                         }
                         else
                         {
-                            await Task.Delay(_cleanupDelay, _dispose.Token)
-                                .ConfigureAwait(false);
+                            await Task.Delay(
+                                _cleanupDelay,
+                                _disposeTokenSource.Token)
+                                    .ConfigureAwait(false);
                         }
                     }
                 }, TaskCreationOptions.LongRunning);
@@ -189,8 +192,8 @@ namespace GreenDonut
                 if (disposing)
                 {
                     Clear();
-                    _dispose?.Cancel();
-                    _dispose?.Dispose();
+                    _disposeTokenSource?.Cancel();
+                    _disposeTokenSource?.Dispose();
                 }
 
                 _disposed = true;